+++ docker-run-codeserver.sh
@@ -23,16 +23,25 @@
 # $1 Volume path
 # $2 Password
 # $3 Proxy domain
+if [ -z "$1" ] || [ -z "$2" ] || [ -z "$3" ] || [ "$1" = "-h" ] || [ "$1" = "--help" ]; then
+    echo "Usage: $0 <volume path> <password> <proxy domain>"
+    echo "EXAMPLE: $0 /docker-volumes/code-server PASSWORD example.com"
+    if [ "$1" = "-h" ] || [ "$1" = "--help" ]; then
+        echo "This script will run code-server on docker"
+        exit 0
+    fi
+    exit 1
+fi
 docker rm -f codeserver
 docker run -d --name=codeserver \
 -p 8377:8443 \
 -e PUID=1026 \
 -e PGID=100 \
 -e TZ=Asia/Riyadh \
--e PASSWORD=$2 \
--e PROXY_DOMAIN=$3 \
--e SUDO_PASSWORD=$2 \
--v $1/config:/config \
+-e PASSWORD="$2" \
+-e PROXY_DOMAIN="$3" \
+-e SUDO_PASSWORD="$2" \
+-v "$1"/config:/config \
 --privileged \
 --restart always \
 ghcr.io/linuxserver/code-server
