+++ glusterfs-disk-wipe.sh
@@ -21,6 +21,15 @@
 # OR
 # curl -fL -H 'Cache-Control: no-cache, no-store' https://raw.githubusercontent.com/alrokayan/scripts/main/glusterfs-disk-wipe.sh | bash -s -- /dev/sdb
 # $1 Disk to wipe, xfs formate, and mount
+if [ -z "$1" ] || [ "$1" = "-h" ] || [ "$1" = "--help" ]; then
+    echo "Usage: $0 <disk to wipe>"
+    echo "EXAMPLE: $0 /dev/sdb"
+    if [ "$1" = "-h" ] || [ "$1" = "--help" ]; then
+        echo "This script will wipe, format, and mount a disk for glusterfs"
+        exit 0
+    fi
+    exit 1
+fi
 DISK=$1
 systemctl stop iptables
 systemctl disable iptables
@@ -29,9 +38,9 @@ apt install xfsprogs glusterfs-server glusterfs-client -y
 systemctl enable glusterd
 systemctl start glusterd
 systemctl status glusterd
-mkfs.xfs $DISK -f
-wipefs -a $DISK
-mkfs.xfs $DISK -f
+mkfs.xfs "$DISK" -f
+wipefs -a "$DISK"
+mkfs.xfs "$DISK" -f
 mkdir /mnt/gfs_disk
 echo "$DISK /mnt/gfs_disk xfs defaults 1 2" >> /etc/fstab
 systemctl daemon-reload
