+++ multipass-delete-all-vms.sh
@@ -20,4 +20,8 @@
 # rm -r scripts && git clone https://github.com/alrokayan/scripts.git && cd scripts && chmod +x * && ./multipass-delete-all-vms.sh
 # OR
 # curl -fL -H 'Cache-Control: no-cache, no-store' https://raw.githubusercontent.com/alrokayan/scripts/main/multipass-delete-all-vms.sh | bash -s
+if [ "$1" = "-h" ] || [ "$1" = "--help" ]; then
+    echo "This script will delete all multipass vms"
+    exit 0
+fi
 multipass delete --purge --all
